requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision
opencv-python>=4.8.0
mediapipe>=0.10.0
pillow>=10.0.0

# Data science
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0

# Backend (for later)
flask>=2.3.0
flask-cors>=4.0.0
